Improve Map canonicalization and add Map iterators.
This CL changes Map storage to use a vector of pairs, which removes
the need to copy the contents twice to sort them and makes it easy to
support Map iteration.  Support for recursive canonicalization is
added as well, and Map::get() uses a binary search when the map is
canonicalized.

Test: cppbor_test_external
Change-Id: Ie7cee5d504e205e1768a26ec5df8436805a6eefe